@@ -21,67 +21,52 @@
     "node": ">=18.15.0"
   },
   "dependencies": {
-    "chalk": "^4.1.0",
+    "chalk": "^5.2.0",
     "lit": "^2.7.5"
   },
   "peerDependencies": {
     "@alaskaairux/design-tokens": "^3.9.0",
     "@alaskaairux/icons": "^4.13.0",
-    "@alaskaairux/webcorestylesheets": "^3.7.3",
-    "@webcomponents/webcomponentsjs": "^2.5.0",
-    "focus-visible": "^5.2.0"
+    "@aurodesignsystem/webcorestylesheets": "^4.6.0"
   },
   "devDependencies": {
-    "@alaskaairux/design-tokens": "3.9.0",
-    "@alaskaairux/icons": "4.13.0",
-    "@alaskaairux/webcorestylesheets": "3.7.3",
-    "@babel/core": "^7.13.14",
-    "@babel/plugin-syntax-dynamic-import": "^7.8.3",
-    "@babel/plugin-transform-runtime": "^7.13.10",
-    "@babel/preset-env": "^7.13.12",
-    "@babel/runtime": "^7.13.10",
-    "@commitlint/cli": "^12.1.1",
-    "@commitlint/config-conventional": "^12.1.1",
-    "@open-wc/testing": "^2.5.32",
-    "@open-wc/testing-karma": "^4.0.9",
-    "@rollup/plugin-node-resolve": "^11.2.1",
+    "@alaskaairux/design-tokens": "^3.14.0",
+    "@alaskaairux/icons": "^4.28.0",
+    "@aurodesignsystem/webcorestylesheets": "^4.6.0",
+    "@commitlint/cli": "^17.6.6",
+    "@commitlint/config-conventional": "^17.6.6",
+    "@open-wc/testing": "^3.2.0",
+    "@rollup/plugin-node-resolve": "^15.1.0",
     "@semantic-release/changelog": "^5.0.1",
     "@semantic-release/git": "^9.0.0",
     "@semantic-release/npm": "^7.1.0",
-    "@web/dev-server": "^0.1.31",
+    "@web/dev-server": "^0.2.3",
     "@web/test-runner": "^0.16.1",
-    "@webcomponents/webcomponentsjs": "^2.5.0",
     "autoprefixer": "^10.2.5",
-    "babel-loader": "^8.2.5",
     "compression": "^1.7.4",
     "concat": "^1.0.3",
+    "concurrently": "^8.2.0",
     "copyfiles": "^2.4.1",
     "core-js": "^3.10.0",
-    "eslint": "^7.23.0",
+    "eslint": "^8.43.0",
     "eslint-plugin-lit": "^1.3.0",
-    "focus-visible": "^5.2.0",
     "husky": "^8.0.3",
     "lodash": "^4.17.21",
-    "marked": "^2.0.1",
     "nodemon": "^2.0.16",
     "npm-run-all": "^4.1.5",
     "postcss": "^8.2.9",
-    "postcss-custom-properties": "^11.0.0",
-    "postcss-discard-comments": "^4.0.2",
+    "postcss-discard-comments": "^6.0.0",
     "postcss-remove-rules": "^1.0.0",
     "postcss-selector-replace": "^1.0.2",
-    "prismjs": "^1.23.0",
     "rollup": "^2.44.0",
     "rollup-plugin-minify-html-literals": "^1.2.6",
     "rollup-plugin-serve": "^1.1.0",
     "semantic-release": "^17.4.2",
-    "sinon": "^10.0.0",
-    "stylelint": "^13.12.0",
-    "stylelint-config-standard": "^21.0.0",
+    "stylelint": "^15.9.0",
+    "stylelint-config-standard": "^33.0.0",
     "typescript": "^5.1.3",
     "wc-sass-render": "^1.2.3",
     "web-component-analyzer": "^1.1.6",
-    "webpack-merge": "^5.7.3",
     "whatwg-fetch": "^3.6.2",
     "yaml-lint": "^1.2.4"
   },
