@@ -1,6 +1,8 @@
 'use strict';
 
-const chalk = require('chalk');
+import chalk from 'chalk';
+import { createRequire } from 'node:module';
+const require = createRequire(import.meta.url);
 const pjson = require('../package.json');
 
 console.log(chalk.hex('#f26135')(`
